Rename files in the "articles/" folder
This is a big rename patch, moving articles across directories.

The point of this patch is to make it use the same structure for
both URLs and files. Without this, we would need to look for a file
in a haystack (i.e. the big folder named "articles") each time.

Links in articles should be modified following the rename, which
should be addressed in the next commit.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>