Writing style.
While I'm not a documentation writer by profession, I am something of a wordsmith and the current state of the documentation makes me cringe.  
I understand if English is not a first language this can be challenging, and am happy to add my 2cs where it might help.  
The intro has run-on sentences, bad grammar, lack of clarity, etc., etc. and so I am submitting this.  
BUT.... please, consider enlisting someone who is a doc author to help.  These docs are currently really rough given the relative maturity of the software.