Fix wrong number of retry times
See also: https://github.com/fluent/fluentd/pull/3649

Before the fix, the number was mistakenly 19.
This fixes it to 18.
(The document was also wrong.)

n-th retry is triggered at `2^n - 1` seconds elapsed since the first flush.
So 18-th retry is at 262143 seconds elapsed and this exceeds 72 hours (259200 seconds).
If the next retry is going to exceed this time limit, the last retry will be
made at exactly this time limit.
So the last time is 18.

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>