Fix default value of rotate_age
It seems that the old value came from `ServerEngine`.

https://github.com/treasure-data/serverengine/blob/v2.3.1/lib/serverengine/daemon_logger.rb#L33

However, actually, `ServerEngine` has nothing to do with the
rotation, and this value `5` is not used at all.

Actually, `Logger::LogDevice` manages the rotation, and the default
value is `7`.

https://github.com/ruby/logger/blob/v1.5.3/lib/logger/log_device.rb#L20

In Fluentd code, the parameters of `Fluent::LogDeviceIO` decides
the behavior of the rotation.

https://github.com/fluent/fluentd/blob/v1.15.3/lib/fluent/supervisor.rb#L563

The parameters of `ServerEngine::DaemonLogger` is meaningless.

https://github.com/fluent/fluentd/blob/v1.15.3/lib/fluent/supervisor.rb#L580-L582

Signed-off-by: Daijiro Fukuda <fukuda@clear-code.com>