Copy missing LICENSE file into 1.0 branch too
LICENSE file was already commit into master branch,
but as it is not default branch, it may be confused
what license is applied to all components.

The documentation of footer contains mention about Apache 2.0
license, but there are some content files which lacks
that notification. (It should be also fixed too)

This situation should be solved by put LICENSE file into
1.0 branch too.

Signed-off-by: Kentaro Hayashi <hayashi@clear-code.com>